+++ ChangeLog.md
@@ -3,10 +3,15 @@
 ### 2018-07-23  Thomas Deuling  <typo3@coding.ms>, Dieter Bruening <dbruening@coding.ms>
 
 *   [TASK] Add constants for feature main menu css-classes.
+*   [FEATURE] Adding simple image content element.
+
+
 
 ### 2018-07-19  Thomas Deuling  <typo3@coding.ms>, Dieter Bruening <dbruening@coding.ms>
 
 *   [BUGFIX] Fixing rendering definition for Divider, Header, Html, Shortcut and Table content element.
+
+
 
 ### 2018-07-18  Thomas Deuling  <typo3@coding.ms>, Dieter Bruening <dbruening@coding.ms>